Fixed trainer form , fixed versioning layout

+++ src/app/shared/course.service.ts
@@ -1,6 +1,9 @@
 import { Injectable } from "@angular/core";
 import { Course } from "./course.model";
 import { HttpClient } from "@angular/common/http";
+
+import { GoogleLoginService } from "../shared/google-login.service";
+
 @Injectable({
   providedIn: "root",
 })
@@ -8,7 +11,10 @@ export class CourseService {
   formData: Course;
 
   list: Course[];
-  constructor(private http: HttpClient) {}
+  constructor(
+    private http: HttpClient,
+    private googleloginservice: GoogleLoginService
+  ) {}
 
   postCourse(formData: Course) {
     return this.http.post("http://localhost:8080/course/post", formData);
@@ -18,6 +24,10 @@ export class CourseService {
   }
 
   updateCourse(formData: Course) {
+    if (!this.googleloginservice.superUser) {
+      this.googleloginservice.checkSuperUser();
+      return;
+    }
     return this.http.put(
       "http://localhost:8080/course/update/" + formData.id,
       formData
@@ -25,6 +35,11 @@ export class CourseService {
   }
 
   deleteCourse(id: number) {
+    if (!this.googleloginservice.superUser) {
+      this.googleloginservice.checkSuperUser();
+      return;
+    }
+
     return this.http.delete("http://localhost:8080/course/delete/" + id);
   }
